requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
shap>=0.42.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
joblib>=1.2.0
drf-spectacular>=0.26.0
pymysql==1.1.1
drf-yasg>=1.20.0